Full intercompatibility between Unix and Windows based systems

+++ Source Code/BuildRKRShell.sh
@@ -16,10 +16,23 @@ then
 	# Compiles using Flex & Bison and the g++ compiler
 	sh FlexBisonWrapper.sh $file $output
 
-	# Removes all useless files
-	rm lex.yy.c
-	rm $file.tab.*
-	rm $logFile
+	# Removes generated Flex file
+	if [ -f "lex.yy.c" ]
+	then
+		rm lex.yy.c
+	fi
+
+	# Removes generated Bison files
+	if [ -f "$file.tab.c" ] && [ -f "$file.tab.h" ]
+	then
+		rm $file.tab.*
+	fi
+
+	# Removes generated RKRShell's log file
+	if [ -f "$logFile" ]
+	then
+		rm $logFile
+	fi
 else
 	# Shows error message
 	printf '\nToo much arguments, expected none\n\n'
